requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.7
multidict==6.0.5
packaging==24.0
pillow==10.3.0
PyJWT==2.8.0
pyngrok==7.1.6
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
twilio==9.0.5
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4
